API

Base

Missing docstring.

Missing docstring for AbstractTrees.ImplicitParents. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AbstractTrees.ImplicitSiblings. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AbstractTrees.ParentLinks. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AbstractTrees.SiblingLinks. Check Documenter's build log for details.

AbstractTrees.StoredParents — Type

Indicates that this tree stores parent links explicitly. The implementation is responsible for defining the parentind function to expose this information.

AbstractTrees.StoredSiblings — Type

Indicates that this tree stores sibling links explicitly, or can compute them quickly (e.g. because the tree has a (small) fixed branching ratio, so the current index of a node can be determined by quick linear search). The implementation is responsible for defining the relative_state function to expose this information.

AbstractTrees.children — Function

children(x)

Get the immediate children of node x.

This is the primary function that needs to be implemented for custom tree types. It should return an iterable object for which an appropriate implementation of Base.pairs is available.

The default behavior is to assume that if an object is iterable, iterating over it gives its children. Non-iterable types are treated as leaf nodes.

AbstractTrees.nodetype — Function

nodetype(tree)

A trait function, defined on the tree object, specifying the types of the nodes. The default is Any. When applicable, define this trait to make iteration inferrable.

Example

struct IntTree
    num::Int
    children::Vector{IntTree}
end
AbstractTrees.children(itree::IntTree) = itree.children
AbstractTrees.nodetype(::IntTree) = IntTree

This suffices to make iteration over, e.g., Leaves(itree::IntTree) inferrable.

Missing docstring.

Missing docstring for AbstractTrees.parentlinks. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AbstractTrees.siblinglinks. Check Documenter's build log for details.

Missing docstring.

Missing docstring for treemap. Check Documenter's build log for details.

Missing docstring.

Missing docstring for treemap!. Check Documenter's build log for details.

Printing

AbstractTrees.DEFAULT_CHARSET — Constant

Default charset argument used by print_tree.

AbstractTrees.ASCII_CHARSET — Constant

Charset using only ASCII characters.

AbstractTrees.TreeCharSet — Type

TreeCharSet

Set of characters (or strings) used to pretty-print tree branches in print_tree.

AbstractTrees.print_tree — Function

print_tree(tree; kwargs...)
print_tree(io::IO, tree; kwargs...)
print_tree(f::Function, io::IO, tree; kwargs...)

Print a text representation of tree to the given io object.

Arguments

f::Function - custom implementation of printnode to use. Should have the signature f(io::IO, node).

maxdepth::Integer = 5 - truncate printing of subtrees at this depth.

indicate_truncation::Bool = true - print a vertical ellipsis character beneath truncated nodes.

charset::TreeCharSet - TreeCharSet to use to print branches.

Examples

julia> tree = [1:3, "foo", [[[4, 5], 6, 7], 8]];

julia> print_tree(tree)
Array{Any,1}
├─ UnitRange{Int64}
│  ├─ 1
│  ├─ 2
│  └─ 3
├─ "foo"
└─ Array{Any,1}
   ├─ Array{Any,1}
   │  ├─ Array{Int64,1}
   │  │  ├─ 4
   │  │  └─ 5
   │  ├─ 6
   │  └─ 7
   └─ 8

julia> print_tree(tree, maxdepth=2)
Array{Any,1}
├─ UnitRange{Int64}
│  ├─ 1
│  ├─ 2
│  └─ 3
├─ "foo"
└─ Array{Any,1}
   ├─ Array{Any,1}
   │  ⋮
   │
   └─ 8

julia> print_tree(tree, charset=AbstractTrees.ASCII_CHARSET)
Array{Any,1}
+-- UnitRange{Int64}
|   +-- 1
|   +-- 2
|   \-- 3
+-- "foo"
\-- Array{Any,1}
    +-- Array{Any,1}
    |   +-- Array{Int64,1}
    |   |   +-- 4
    |   |   \-- 5
    |   +-- 6
    |   \-- 7
    \-- 8

AbstractTrees.printnode — Function

printnode(io::IO, node)

Print a compact representation of a single node.

AbstractTrees.repr_node — Function

repr_node(node; context=nothing)

Get the string representation of a node using printnode. This works analagously to Base.repr.

context is an IO or IOContext object whose attributes are used for the I/O stream passed to printnode.

AbstractTrees.repr_tree — Function

repr_tree(tree; context=nothing, kw...)

Get the string result of calling print_tree with the supplied arguments.

The context argument works as it does in Base.repr.

Iteration

AbstractTrees.Leaves — Type

Iterator to visit the leaves of a tree, e.g. for the tree

Any[1,Any[2,3]]
├─ 1
└─ Any[2,3]
   ├─ 2
   └─ 3

we will get [1,2,3].

AbstractTrees.PostOrderDFS — Type

Iterator to visit the nodes of a tree, guaranteeing that children will be visited before their parents.

e.g. for the tree

Any[1,Any[2,3]]
├─ 1
└─ Any[2,3]
   ├─ 2
   └─ 3

we will get [1, 2, 3, [2, 3], [1, [2, 3]]].

AbstractTrees.PreOrderDFS — Type

Iterator to visit the nodes of a tree, guaranteeing that parents will be visited before their children.

Optionally takes a filter function that determines whether the iterator should continue iterating over a node's children (if it has any) or should consider that node a leaf.

e.g. for the tree

Any[Any[1,2],Any[3,4]]
├─ Any[1,2]
|  ├─ 1
|  └─ 2
└─ Any[3,4]
   ├─ 3
   └─ 4

we will get [[[1, 2], [3, 4]], [1, 2], 1, 2, [3, 4], 3, 4].

Invalidation

Modifying the underlying tree while iterating over it, is allowed, however, if parents and sibling links are not explicitly stored, the identity of any parent of the last obtained node does not change (i.e. mutation is allowed, replacing nodes is not).

AbstractTrees.StatelessBFS — Type

Iterator to visit the nodes of a tree, all nodes of a level will be visited before their children

e.g. for the tree

Any[1,Any[2,3]]
├─ 1
└─ Any[2,3]
   ├─ 2
   └─ 3

we will get [[1, [2,3]], 1, [2, 3], 2, 3].

WARNING: This is $O(n^2)$, only use this if you know you need it, as opposed to a more standard statefull approach.

Missing docstring.

Missing docstring for TreeIterator. Check Documenter's build log for details.

Other

Missing docstring.

Missing docstring for AnnotationNode. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ShadowTree. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Tree. Check Documenter's build log for details.